hacer componente home
